step

+++ new6/math_test.cc
@@ -24,7 +24,7 @@ template <class D> void TestMath(D d, uint64_t start, uint64_t stop) {
 } // namespace hwy
 int main(int argc, char *argv[]) {
 	(void)argc;
-  hwy::N_EMU128::Simd<double, 1, 0> b2;
+  hwy::N_EMU128::Simd<long double, 1, 0> b2;
   uint64_t start =  std::stoull(argv[1]);
   uint64_t stop = std::stoull (argv[2]);
   hwy::TestMath(b2, start, stop);
